Flask/requirements.txt
Flask==2.1.3
nltk==3.7
numpy==1.21.6
torch
gunicorn
SpeechRecognition==3.8.1
urllib3==1.26.11
cloudinary==1.29.0
pyttsx3==2.71
python-dotenv==0.20.0